+++ assets/js/workflowManualSelectionQuickEdit.js
@@ -31,6 +31,7 @@ function Fieldset(_ref) {
     apiUrl = _ref.apiUrl,
     nonce = _ref.nonce,
     workflowNonce = _ref.workflowNonce,
+    actionNonce = _ref.actionNonce,
     onChange = _ref.onChange,
     wrapper = _ref.wrapper;
   var _useSelect = (0,_wordpress_data__WEBPACK_IMPORTED_MODULE_0__.useSelect)(function (select) {
@@ -95,6 +96,10 @@ function Fieldset(_ref) {
     type: "hidden",
     name: "future_workflow_view",
     value: context
+  }), context === 'quick-edit' && actionNonce && /*#__PURE__*/React.createElement("input", {
+    type: "hidden",
+    name: "_future_action_nonce",
+    value: actionNonce
   }), controls));
 }
 
@@ -2189,12 +2194,14 @@ window.inlineEditPost.edit = function (button, id) {
   var apiUrl = window.futureWorkflowManualSelection.apiUrl;
   var nonce = window.futureWorkflowManualSelection.nonce;
   var workflowNonce = window.futureWorkflowManualSelection.workflowNonce;
+  var actionNonce = window.futureWorkflowManualSelection.actionNonce;
   var component = /*#__PURE__*/React.createElement(_fieldset__WEBPACK_IMPORTED_MODULE_1__.Fieldset, {
     context: "quick-edit",
     postId: postId,
     apiUrl: apiUrl,
     nonce: nonce,
     workflowNonce: workflowNonce,
+    actionNonce: actionNonce,
     wrapper: function wrapper(_ref) {
       var workflowsWithManualTrigger = _ref.workflowsWithManualTrigger,
         children = _ref.children;